move firebase related ops to firebase.js
